Make core time arithmetic width-deterministic (uint32_t)
millis() is 32 bits wide on the target: the wrap at 2^32 ms IS the
semantics of every timestamp comparison. Declaring time values as
unsigned long makes that width platform-dependent - correct on the
ESP32 where unsigned long is 32-bit, silently different on the 64-bit
host test build, where the wrap never occurs. Consequences: wrap
behavior was untestable (see the discussion on the wrap-safety PR),
and the suite exercised different arithmetic than the target runs.

This narrows the core time-carrying variables to uint32_t:

- Software.cpp: currentMillis, the interval trackers, and the
  update_calculated_values() parameter
- comm_contactorcontrol.cpp: precharge/BMS-reset state machine
  timestamps and the reset-interval helper
- datalayer: remote_set_timestamp / remote_set_timeout
- MyTimer: interval and previous_millis

and moves the remote-limit expiry check into
update_remote_limit_expiry() in safety.cpp so it is reachable from the
host test suite.

With production widths deterministic, the emulated millis() now
truncates to 32 bits exactly like the target, so the host runs the
same arithmetic the ESP32 runs - wrap included. The BMS-reset
overflow tests move from the host's 2^64 boundary to the target's real
2^32 boundary, and new wrap regression tests cover the two real bugs
fixed by the wrap-safety sweep: the SMA startup-grace re-arm and the
remote-limit expiry (both directions of the old overflow).

Driver-level unsigned long usage (including the transmit_can interface
signatures) is deliberately untouched - that sweep is a follow-up.

Depends on the wrap-safety branch (built on top of it).

+++ Software/src/communication/contactorcontrol/comm_contactorcontrol.cpp
@@ -41,16 +41,16 @@ uint16_t pwm_hold_duty = 250;
 #define PWM_OFF_DUTY 0  //No need to have this userconfigurable
 #define PWM_Positive_Channel 0
 #define PWM_Negative_Channel 1
-static unsigned long prechargeStartTime = 0;
-unsigned long negativeStartTime = 0;
-unsigned long prechargeCompletedTime = 0;
-unsigned long timeSpentInFaultedMode = 0;
-unsigned long currentTime = 0;
-unsigned long lastPowerRemovalTime = 0;
+static uint32_t prechargeStartTime = 0;
+uint32_t negativeStartTime = 0;
+uint32_t prechargeCompletedTime = 0;
+uint32_t timeSpentInFaultedMode = 0;
+uint32_t currentTime = 0;
+uint32_t lastPowerRemovalTime = 0;
 bool periodicResetDeferred = false;   //True while a due periodic reset is waiting for SOC to recover
 bool balancingPeriodSkipped = false;  //True once balancing has cost the reset a period
-unsigned long bmsPowerOnTime = 0;
-const unsigned long bmsWarmupDuration = 3000;
+uint32_t bmsPowerOnTime = 0;
+const uint32_t bmsWarmupDuration = 3000;
 #define BMS_RESET_DEFER_SOC_PPTT 1500  // 15.00%, below this the low-SOC guard defers the periodic reset
 
 void set(uint8_t pin, bool direction, uint32_t pwm_freq = 0xFFFF) {
@@ -337,12 +337,12 @@ void bms_power_on() {
 
 // Configured period between two automatic resets. Guarded to 24h in case the stored
 // value is missing or nonsensical, see load_settings().
-static unsigned long bms_reset_interval_ms() {
+static uint32_t bms_reset_interval_ms() {
   uint32_t hours = periodic_bms_reset_interval_h;
   if (hours == 0) {
     hours = 24;
   }
-  return (unsigned long)hours * 60UL * 60UL * 1000UL;
+  return (uint32_t)hours * 60UL * 60UL * 1000UL;
 }
 
 /* The two guards behave differently on purpose, so the decision is three-way rather than
